LayoutTests/fast/forms/datalist/datalist-child-validation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<datalist id="list">
  <input type=text id=e required>
</datalist>

<script>
description('Test for child elements of a datalist element.');

var e = document.getElementById('e');
shouldBeFalse('e.willValidate');
document.body.appendChild(e);
shouldBeTrue('e.willValidate');
shouldBe('document.querySelector(":invalid")', 'e');

</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
