LayoutTests/fast/forms/datalist/datalist-child-validation-expected.txt

Test for child elements of a datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.willValidate is false
PASS e.willValidate is true
PASS document.querySelector(":invalid") is e
PASS successfullyParsed is true

TEST COMPLETE
